Controller Implemented

+++ parkinglotjava/parkinglot/DTO/IssueTicketRequest.java
@@ -6,7 +6,7 @@ public class IssueTicketRequest {
     
     private String vehicleNo;
 
-    private long gateId;
+    private int gateId;
 
     private String ownerName;
 
@@ -22,11 +22,11 @@ public class IssueTicketRequest {
         this.vehicleNo = vehicleNo;
     }
 
-    public long getGateId() {
+    public int getGateId() {
         return this.gateId;
     }
 
-    public void setGateId(long gateId) {
+    public void setGateId(int gateId) {
         this.gateId = gateId;
     }
